:mag: Move GameStateProvider out of server
